+++ scripts/utils/n3rMotionPose_tools.py
@@ -1050,7 +1050,49 @@ def draw_wind_icon(img, wind_delta, pos=(50,50), scale=100, color=(0,255,255), t
 
     return img
 
-#--------------------------------------------------------------------------------------------
+#-------------------------------------------- Micro moton ------------------------------------------------
+
+def apply_micro_motion(
+    latents: torch.Tensor,
+    frame_counter: int,
+    device,
+    masks: dict,
+    strength: float = 0.3,   # 🔥 NOUVEAU 0.3 - 0.6 → très réaliste (cinéma) - stable 0.25
+    randomize: bool = True
+):
+    """
+    Micro motion avec contrôle global de l'intensité.
+    strength : 0 = OFF, 1 = normal, >1 = amplifié
+    """
+
+    # 🔹 Clamp sécurité (évite explosion)
+    strength = max(0.0, min(strength, 5.0))
+
+    t = torch.tensor(frame_counter / 6.0, device=device)
+
+    for zone_name, params in masks.items():
+        if params is None:
+            continue
+
+        mask, phase, amplitude = params
+        if mask is None:
+            continue
+
+        mask = mask.to(dtype=latents.dtype, device=device)
+
+        # 🔹 Micro random plus stable (moins agressif)
+        if randomize:
+            noise = (torch.rand_like(mask) - 0.5) * 0.01
+        else:
+            noise = 0.0
+
+        # 🔥 Application du strength AU BON ENDROIT
+        delta = strength * amplitude * mask * torch.sin(t + phase + noise)
+
+        latents = latents + delta
+
+    return latents
+#-------------------------------------------- Gestion du vent ------------------------------------------------
 
 def apply_hair_motion_cycle(
     latents,
@@ -1062,6 +1104,7 @@ def apply_hair_motion_cycle(
     device,
     delta_px=None,
     prev_hair_field=None,
+    strength=0.5,
     debug=False,
     debug_dir=None
 ):
@@ -1072,7 +1115,7 @@ def apply_hair_motion_cycle(
     2 → apply_hair_motion_extreme
     """
 
-    mode = frame_counter % 3  # cycle 0,1,2
+    mode = frame_counter % 4  # cycle 0,1,2,3
 
     if mode == 0:
         latents_hair, hair_delta = apply_hair_motion_vent(
@@ -1080,7 +1123,7 @@ def apply_hair_motion_cycle(
             frame_counter, device,
             delta_px=delta_px,
             prev_hair_field=prev_hair_field,
-            strength=0.5,
+            strength=strength,
             debug=debug,
             debug_dir=debug_dir
         )
@@ -1090,17 +1133,27 @@ def apply_hair_motion_cycle(
             frame_counter, device,
             delta_px=delta_px,
             prev_hair_field=prev_hair_field,
-            strength=0.5,
+            strength=strength,
             debug=debug,
             debug_dir=debug_dir
         )
-    else:  # mode == 2
+    elif mode == 2:
+        latents_hair, hair_delta = apply_hair_motion_cinema(
+            latents, mask_hair, grid, H, W,
+            frame_counter, device,
+            delta_px=delta_px,
+            prev_hair_field=prev_hair_field,
+            strength=strength,
+            debug=debug,
+            debug_dir=debug_dir
+        )
+    else:  # mode == 3
         latents_hair, hair_delta = apply_hair_motion_extreme(
             latents, mask_hair, grid, H, W,
             frame_counter, device,
             delta_px=delta_px,
             prev_hair_field=prev_hair_field,
-            strength=0.5,
+            strength=strength,
             debug=debug,
             debug_dir=debug_dir
         )
@@ -1371,6 +1424,114 @@ def apply_hair_motion_vent(
     # 🔹 Debug
     if debug:
         print(f"[DEBUG] Hair motion Vent applied with strength={strength:.2f}")
+
+    if debug and debug_dir is not None:
+        debug_save_mask_and_wind(
+            mask=mask_hair,
+            wind_delta=wind_delta,
+            H=H,
+            W=W,
+            debug_dir=debug_dir,
+            frame_counter=frame_counter
+        )
+
+    return latents_out, hair_delta_field
+#------------ version cinema -----------------
+def apply_hair_motion_cinema(
+    latents,
+    mask_hair,
+    grid,
+    H,
+    W,
+    frame_counter,
+    device,
+    delta_px=None,
+    prev_hair_field=None,
+    strength: float = 1.0,   # 🔥 NOUVEAU
+    debug=False,
+    debug_dir=None
+):
+    B = latents.shape[0]
+
+    # 🔹 Clamp sécurité
+    strength = max(0.0, min(strength, 5.0))
+
+    # -------------------- Temps --------------------
+    t = frame_counter
+    t_wind1 = torch.tensor(t / 15.0, device=device)
+    t_wind2 = torch.tensor(t / 60.0, device=device)
+
+    # -------------------- Multi-échelle bruit --------------------
+    def multi_noise(grid, t, scales=[0.05,0.15,0.3], weights=[1.0,0.5,0.25]):
+        val = 0
+        for s,w in zip(scales, weights):
+            val += w * smooth_noise(grid, t, scale=s)
+        return val
+
+    noise_x = multi_noise(grid, t)
+    noise_y = multi_noise(grid, t + 123, scales=[0.08,0.2,0.4], weights=[1.0,0.5,0.25])
+
+    # -------------------- Champ delta de base --------------------
+    hair_delta_field = torch.zeros_like(grid)
+    hair_delta_field[...,0] = 0.03 * noise_x
+    hair_delta_field[...,1] = 0.05 * noise_y
+
+    # -------------------- Vent dynamique --------------------
+    wind_dir = torch.tensor([[1.0,0.2],[0.3,0.1]], device=device).mean(dim=0).view(1,1,1,2)
+    wind_strength = 0.02 + 0.01 * torch.sin(t_wind1) + 0.005 * torch.sin(t_wind2)
+    wind_delta = wind_dir * wind_strength
+
+    # -------------------- Gravité --------------------
+    gravity_delta = torch.zeros_like(grid)
+    gravity_delta[...,1] = 0.004
+
+    # -------------------- Influence du torse --------------------
+    if delta_px is not None:
+        speed = torch.norm(delta_px, dim=-1, keepdim=True)
+        hair_delta_field *= (1.0 + 2.5 * speed)
+        wind_delta *= (1.0 + 1.5 * speed)
+        gravity_delta *= (1.0 + 0.5 * speed)
+
+    # -------------------- Inertie --------------------
+    inertia = 0.85
+    if prev_hair_field is not None:
+        hair_delta_field = inertia * prev_hair_field + (1 - inertia) * hair_delta_field
+
+    # -------------------- Masque + falloff --------------------
+    mask_hair_expand = mask_hair.permute(0,2,3,1)
+    yy = torch.linspace(0,1,H,device=device).view(1,H,1,1)
+    smooth_falloff = yy**2 * (3 - 2*yy)
+    mask_hair_expand = mask_hair_expand * smooth_falloff
+
+    # -------------------- Micro-souplesse --------------------
+    spring = 0.003 * torch.sin(t*0.5 + grid[...,1:2]*3.0)
+    hair_delta_field[...,1:2] += spring
+
+    # -------------------- Micro noise --------------------
+    micro_noise = 0.001 * (torch.rand_like(hair_delta_field)-0.5)
+    hair_delta_field += micro_noise
+
+    # =========================
+    # 🔥 APPLICATION DU STRENGTH (AU BON ENDROIT)
+    # =========================
+    hair_delta_field = hair_delta_field * strength
+    wind_delta = wind_delta * strength
+    gravity_delta = gravity_delta * strength
+
+    # -------------------- Application --------------------
+    grid_hair = grid + hair_delta_field * mask_hair_expand
+    grid_hair += wind_delta * mask_hair_expand
+    grid_hair += gravity_delta * mask_hair_expand
+
+    # -------------------- Normalisation --------------------
+    grid_hair[...,0] = 2.0 * grid_hair[...,0] / (W-1) - 1.0
+    grid_hair[...,1] = 2.0 * grid_hair[...,1] / (H-1) - 1.0
+
+    # -------------------- Sampling --------------------
+    latents_out = F.grid_sample(latents, grid_hair, align_corners=True)
+
+    if debug:
+        print(f"[DEBUG] Hair motion cinema applied | strength={strength:.2f}")
 
     if debug and debug_dir is not None:
         debug_save_mask_and_wind(
